Repository reorganization per .ai/plans/repo-reorganization.md.
- Split main.py into pr_review/ package (13 modules); main.py is a thin
  facade re-exporting Cloud Function entry points.
- Move documentation to docs/ (SETUP, AUTHENTICATION, IAM, etc.).
- Move deployment and setup scripts to scripts/ (deploy.sh, setup-env.sh,
  env-template.sh).
- Move tests to tests/; update mocks from main.* to pr_review.*.
- Update terraform/functions.tf (pr_review/ in archive), scripts/deploy.sh
  (repo root and pr_review/ checks), .gcloudignore (exclude scripts/tests/docs).
- Update .cursorrules, .gitignore, README.md for new layout.
- Add .ai/plans/repo-reorganization.md and .gcloudignore.

No behavior or business logic changes; all entry points remain importable
from main.py for Cloud Functions Gen2.
