+++ azure/azure-pr-pipeline.yml
@@ -16,3 +16,4 @@ steps:
         -H "Content-Type: application/json" \
         -d '{"pr_id": $(System.PullRequest.PullRequestId)}' \
         https://us-central1-YOUR_PROJECT.cloudfunctions.net/pr-regression-review
+
