+++ docs/AUTHENTICATION.md
