+++ pr_review/__init__.py
@@ -1,0 +1,66 @@
+"""PR Review - Automated PR regression analysis package."""
+
+from pr_review.azure_client import AzureDevOpsClient
+from pr_review.config import (
+    DEFAULT_EXTENSIVE_PR_FILE_THRESHOLD,
+    DEFAULT_EXTENSIVE_PR_SIZE_THRESHOLD,
+    load_config,
+    load_webhook_config,
+)
+from pr_review.entry_points import (
+    process_dead_letter_queue,
+    receive_webhook,
+    review_pr,
+    review_pr_pubsub,
+)
+from pr_review.filtering import (
+    FILTERED_FILE_EXTENSIONS,
+    filter_and_limit_files,
+    filter_non_code_files,
+    is_extensive_pr,
+)
+from pr_review.gemini import call_gemini
+from pr_review.idempotency import (
+    MAX_RETRY_ATTEMPTS,
+    check_and_claim_processing,
+    update_marker_completed,
+    update_marker_failed,
+    update_marker_for_retry,
+)
+from pr_review.models import FilterResult, ReviewResult
+from pr_review.prompt import SYSTEM_PROMPT, build_review_prompt
+from pr_review.review import process_pr_review
+from pr_review.severity import get_max_severity
+from pr_review.storage import save_to_storage
+from pr_review.utils import make_response, timed_operation
+
+__all__ = [
+    "AzureDevOpsClient",
+    "DEFAULT_EXTENSIVE_PR_FILE_THRESHOLD",
+    "DEFAULT_EXTENSIVE_PR_SIZE_THRESHOLD",
+    "FILTERED_FILE_EXTENSIONS",
+    "FilterResult",
+    "MAX_RETRY_ATTEMPTS",
+    "ReviewResult",
+    "SYSTEM_PROMPT",
+    "build_review_prompt",
+    "call_gemini",
+    "check_and_claim_processing",
+    "filter_and_limit_files",
+    "filter_non_code_files",
+    "is_extensive_pr",
+    "load_config",
+    "load_webhook_config",
+    "make_response",
+    "process_dead_letter_queue",
+    "process_pr_review",
+    "receive_webhook",
+    "review_pr",
+    "review_pr_pubsub",
+    "save_to_storage",
+    "get_max_severity",
+    "timed_operation",
+    "update_marker_completed",
+    "update_marker_failed",
+    "update_marker_for_retry",
+]
