+++ pr_review/storage.py
@@ -1,0 +1,48 @@
+"""Cloud Storage operations for saving reviews."""
+
+import logging
+from datetime import datetime, timezone
+
+from google.cloud import storage
+
+from pr_review.utils import timed_operation
+
+logger = logging.getLogger("pr_review")
+
+
+def save_to_storage(bucket_name: str, pr_id: int, review: str) -> str:
+    """Save review to Cloud Storage with date partitioning.
+
+    Args:
+        bucket_name: GCS bucket name
+        pr_id: Pull request ID
+        review: Markdown review content
+
+    Returns:
+        Full GCS path (gs://bucket/path)
+    """
+    logger.info(f"[GCS] Saving review for PR #{pr_id} to bucket: {bucket_name}")
+    logger.debug(f"[GCS] Review content size: {len(review)} chars")
+
+    with timed_operation() as elapsed:
+        try:
+            client = storage.Client()
+            bucket = client.bucket(bucket_name)
+
+            # Date partitioning: yyyy/mm/dd
+            now = datetime.now(timezone.utc)
+            date_path = now.strftime("%Y/%m/%d")
+            timestamp = now.strftime("%H%M%S")
+
+            blob_path = f"reviews/{date_path}/pr-{pr_id}-{timestamp}-review.md"
+            blob = bucket.blob(blob_path)
+
+            blob.upload_from_string(review, content_type="text/markdown")
+
+            full_path = f"gs://{bucket_name}/{blob_path}"
+            logger.info(f"[GCS] Upload complete: {blob_path} | {len(review)} bytes | {elapsed():.0f}ms")
+
+            return full_path
+        except Exception as e:
+            logger.error(f"[GCS] Upload FAILED | {elapsed():.0f}ms | Error: {str(e)}")
+            raise
