+++ scripts/env-template.sh
