+++ scripts/setup-env.sh
