+++ tests/test_vertex.py
